7 juli 2023 08:13

+++ database/migrations/2023_07_06_110025_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('status');
             $table->integer('stok');
-            $table->float('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
